+++ Changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Re-design sqlite3 middleware and log in `engine/sqlmanager/src/sqlite3.js`. Completed on 2025-07-17
 - Release a simple design of the `atomic/atom/startupinit` module to help all components initialize the startup mode, especially the database sqlite. Completed on 2025-07-17.
 - Add get_repos_rawfile function to giteapi molecule at atomic which able retrive raw file from gitea tag in json format. Completed on 2025-07-31
+- Apply session.destroy function into deskelectronjs engine. Completed on 2025-08-07
 
 ### Changed