+++ Changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Release a simple design of the `atomic/atom/startupinit` module to help all components initialize the startup mode, especially the database sqlite. Completed on 2025-07-17.
 - Add get_repos_rawfile function to giteapi molecule at atomic which able retrive raw file from gitea tag in json format. Completed on 2025-07-31
 - Apply session.destroy function into deskelectronjs engine. Completed on 2025-08-07
+- Add missing and upgrade package. Completed on 2025-08-011
 
 ### Changed