+++ Changelog.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2025-08-13
+
+### Summary
+
+### Added
+
+- Upgrade hono version to 4.9.1 for both webbunjs and webnodehonojs engine. Completed on 2025-08-013
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+[1.0.4]: https://github.com/wkloh76/oricommjs_v2/releases/tag/1.0.4
+
 ## [1.0.3] - 2025-07-17
 
 ### Summary